requirements/LOCK_qs-s_freeze.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.0
attrs==25.4.0
Authlib==1.6.6
autopep8==2.3.2
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
confluent-kafka==2.13.0
cryptography==46.0.3
dill==0.4.0
Flask==3.1.2
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.3.0
MarkupSafe==3.0.3
multidict==6.7.0
multiprocess==0.70.18
pip==25.3
propcache==0.4.1
psutil==7.2.0
pycodestyle==2.14.0
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyzmq==27.1.0
redis==7.1.0
requests==2.32.5
schwabdev==3.0.0
schwab-py==1.5.1
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
websockets==15.0.1
Werkzeug==3.1.4
wheel==0.45.1
yarl==1.22.0
